Tidied up code.
Changed iva.h so that it is in iva.c (as it should be) then created an
appropriate header file. Likewise with additional_math.h (now called
additional_math.c with a proper header file).

Minor comments added to the loop break in the IVA code. Basically it's a
comparison of the change of a cost function value, if it's smaller than a
value it breaks from the loop.